feat: move preview script to new command in Hydrogen CLI
