+++ examples/template-hydrogen-default/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "hydrogen-template-default",
+  "name": "hydrogen-app",
   "description": "This the default template for Hydrogen",
   "version": "0.9.0",
   "license": "MIT",
@@ -14,13 +14,11 @@
     "build:server": "vite build --outDir dist/server --ssr src/entry-server.jsx",
     "build:worker": "cross-env WORKER=true vite build --outDir dist/worker --ssr worker.js",
     "serve": "node --enable-source-maps server",
-    "preview": "node ./scripts/preview.mjs"
+    "preview": "hydrogen preview"
   },
   "prettier": "@shopify/prettier-config",
   "devDependencies": {
-    "@miniflare/core": "2.2.0",
-    "@miniflare/http-server": "2.2.0",
-    "@miniflare/runner-vm": "2.2.0",
+    "@shopify/hydrogen-cli": "^0.9.0",
     "@shopify/prettier-config": "^1.1.2",
     "@shopify/stylelint-plugin": "^10.0.1",
     "@tailwindcss/typography": "^0.5.0",
@@ -33,9 +31,7 @@
     "prettier": "^2.3.2",
     "stylelint": "^13.13.0",
     "tailwindcss": "^3.0.0",
-    "vite": "^2.7.0",
-    "connect": "^3.7.0",
-    "mime": "^3.0.0"
+    "vite": "^2.7.0"
   },
   "dependencies": {
     "@headlessui/react": "^1.4.1",
