+++ packages/cli/package.json
@@ -44,16 +44,23 @@
     "prettier": "^2.4.1"
   },
   "devDependencies": {
+    "@miniflare/core": "^2.2.0",
+    "@miniflare/runner-vm": "^2.2.0",
+    "@miniflare/shared": "^2.2.0",
+    "@miniflare/web-sockets": "^2.2.0",
     "@types/change-case": "^2.3.1",
     "@types/faker": "^5.5.7",
     "@types/fs-extra": "^9.0.12",
     "@types/inquirer": "^7.3.3",
+    "@types/mime": "^2.0.3",
     "@types/minimist": "^1.2.2",
     "@types/semver": "^7.3.9",
     "change-case": "^4.1.2",
+    "connect": "^3.7.0",
     "faker": "^5.5.3",
     "get-port": "^5.1.1",
     "inquirer": "^8.1.2",
+    "mime": "^3.0.0",
     "playwright-chromium": "^1.13.0",
     "semver": "^7.3.5",
     "sirv": "^1.0.14",
