+++ packages/cli/src/commands/preview/mini-oxygen/storage.ts
@@ -1,9 +1,11 @@
 export class StorageFactory {
+  storages: Map<string, any>;
+
   constructor() {
     this.storages = new Map();
   }
 
-  storage(namespace) {
+  storage(namespace: string) {
     let storage = this.storages.get(namespace);
     if (storage) return storage;
     this.storages.set(namespace, (storage = new Map()));
